migrate editor to svelte 4 and sveltekit 2
