+++ packages/editor/__tests__/setupVitest.ts
@@ -1,0 +1,79 @@
+/**
+ * Based on: https://github.com/davipon/svelte-component-test-recipes/blob/fc493cd3076404dc92e74dc71b67c0e8be022af1/setupTest.ts
+ */
+
+/* eslint-disable @typescript-eslint/no-empty-function */
+import * as matchers from "@testing-library/jest-dom/matchers"; 
+import { expect, vi } from 'vitest';
+import type { Navigation, Page } from '@sveltejs/kit';
+import { readable } from 'svelte/store';
+import * as environment from '$app/environment';
+import * as navigation from '$app/navigation';
+import * as stores from '$app/stores';
+
+expect.extend(matchers);
+
+// Mock SvelteKit runtime module $app/environment
+vi.mock('$app/environment', (): typeof environment => ({
+	browser: false,
+	dev: true,
+	building: false,
+	version: 'any'
+}));
+
+// Mock SvelteKit runtime module $app/navigation
+vi.mock('$app/navigation', (): typeof navigation => ({
+	afterNavigate: () => {},
+	beforeNavigate: () => {},
+	disableScrollHandling: () => {},
+	goto: () => Promise.resolve(),
+	invalidate: () => Promise.resolve(),
+	invalidateAll: () => Promise.resolve(),
+	preloadData: () => Promise.resolve(),
+	preloadCode: () => Promise.resolve()
+}));
+
+// Mock SvelteKit runtime module $app/stores
+vi.mock('$app/stores', (): typeof stores => {
+	const getStores: typeof stores.getStores = () => {
+		const navigating = readable<Navigation | null>(null);
+		const page = readable<Page>({
+			url: new URL('http://localhost'),
+			params: {},
+			route: {
+			  id: null
+			},
+			status: 200,
+			error: null,
+			data: {},
+			form: undefined
+		});
+		const updated = { subscribe: readable(false).subscribe, check: async() => false };
+
+		return { navigating, page, updated };
+	};
+
+	const page: typeof stores.page = {
+		subscribe(fn) {
+			return getStores().page.subscribe(fn);
+		}
+	};
+	const navigating: typeof stores.navigating = {
+		subscribe(fn) {
+			return getStores().navigating.subscribe(fn);
+		}
+	};
+	const updated: typeof stores.updated = {
+		subscribe(fn) {
+			return getStores().updated.subscribe(fn);
+		},
+		check: async () => false
+	};
+
+	return {
+		getStores,
+		navigating,
+		page,
+		updated
+	};
+});
