+++ packages/editor/package.json
@@ -53,18 +53,19 @@
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@playwright/test": "^1.32.1",
-    "@sveltejs/adapter-node": "^1.2.0",
-    "@sveltejs/kit": "^1.15.2",
-    "@testing-library/dom": "^9.2.0",
-    "@testing-library/jest-dom": "^5.16.5",
+    "@sveltejs/adapter-node": "^2.0.0",
+    "@sveltejs/kit": "^2.0.0",
+    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@testing-library/dom": "^9.3.4",
+    "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/svelte": "^3.2.2",
-    "@testing-library/user-event": "^14.4.3",
+    "@testing-library/user-event": "^14.5.2",
     "@tsconfig/recommended": "^1.0.2",
     "@types/node": "^18.15.11",
-    "@types/testing-library__jest-dom": "^5.14.5",
+    "@types/testing-library__jest-dom": "^6.0.0",
     "@videobrew/client": "^0.4.5",
     "@vitest/coverage-c8": "0.30.1",
-    "@vitest/coverage-istanbul": "0.30.1",
+    "@vitest/coverage-v8": "^1.3.1",
     "autoprefixer": "^10.4.13",
     "concurrently": "8.0.1",
     "cross-env": "^7.0.3",
@@ -73,17 +74,17 @@
     "nyc": "^15.1.0",
     "postcss": "^8.4.21",
     "source-map-support": "^0.5.21",
-    "svelte": "^3.54.0",
-    "svelte-check": "^3.0.1",
+    "svelte": "^4.0.0",
+    "svelte-check": "^3.4.3",
     "svelte-fragment-component": "^1.2.0",
     "svelte-htm": "^1.2.0",
     "tailwindcss": "^3.2.7",
     "ts-node": "^10.9.1",
     "tslib": "^2.4.1",
     "typescript": "5.0.4",
-    "vite": "^4.0.0",
-    "vite-plugin-istanbul": "^4.0.1",
-    "vitest": "0.30.1",
+    "vite": "^5.1.4",
+    "vite-plugin-istanbul": "^4.1.0",
+    "vitest": "^1.3.1",
     "wait-on": "^7.0.1",
     "ws": "^8.13.0"
   },
