+++ packages/editor/vite.config.ts
@@ -1,11 +1,6 @@
-import istanbul from 'vite-plugin-istanbul';
 import { sveltekit } from '@sveltejs/kit/vite';
 import type { UserConfig } from 'vite';
 import { configDefaults, type UserConfig as VitestConfig } from 'vitest/config';
-import { exec } from 'child_process';
-import util from 'util';
-import fs from 'fs';
-import path from 'path';
 
 const config: UserConfig & { test: VitestConfig['test'] } = {
 	build: {
@@ -13,10 +8,6 @@ const config: UserConfig & { test: VitestConfig['test'] } = {
 	},
 	plugins: [
 		sveltekit(),
-		istanbul({
-			nycrcPath: path.resolve(__dirname, '.nycrc'),
-      requireEnv: true,
-		}),
 	],
 	server: {
 		port: 8087,
@@ -26,15 +17,12 @@ const config: UserConfig & { test: VitestConfig['test'] } = {
     'import.meta.vitest': 'undefined'
   },
 	test: {
-    globals: true,
     environment: 'jsdom',
 		includeSource: [
 			'src/**/*.{js,ts,svelte}'
 		],
-    // Add @testing-library/jest-dom matchers & mocks of SvelteKit modules
-    setupFiles: ['__tests__/setup-vitest.ts'],
+    setupFiles: ['__tests__/setupVitest.ts'],
 		coverage: {
-			provider: 'istanbul',
 			reportsDirectory: 'coverage/component',
 			all: true,
 			include: ['src/**/*.{js,ts,svelte}'],
@@ -43,7 +31,7 @@ const config: UserConfig & { test: VitestConfig['test'] } = {
 		exclude: [
 			...configDefaults.exclude,
 			// Exclude playwright tests folder
-			'__tests__/e2e'
+			'__tests__/e2e',
 		]
 	}
 };
